geocoder v0.7.0 Geocoder.Data protocol

Types

t :: term

Functions

address(data)

The street address that you want to geocode, in the format used by the national postal service of the country concerned. Additional address elements such as business names and unit, suite or floor numbers should be avoided. Please refer to the FAQ for additional guidance.

@tags [direct] @see components

bounds(data)

The bounding box of the viewport within which to bias geocode results more prominently. This parameter will only influence, not fully restrict, results from the geocoder. (For more information see Viewport Biasing.)

@tags [direct]

components(data)

A component filter for which you wish to obtain a geocode. See Component Filtering for more information.

The components filter will also be accepted as an optional parameter if an address is provided.

@tags [direct]

endpoint(data, type)

Specifies the endpoint for the given implementation.

address: simple string, designationg an address, or it’s part.

components: the component filters, separated by a pipe (|). Each component filter consists of a component:value pair and will fully restrict the results from the geocoder.

place_id: The place ID is a unique identifier that can be used with other Google APIs. For example, you can use the placeID returned by the Google Maps Roads API to get the address for a snapped point. For more information about place IDs, see the place ID overview. The place ID may only be specified if the request includes an API key or a Google Maps APIs Premium Plan client ID.

result_type: examples of address types: country, street_address, postal_code. For a full list of allowable values, see the address types. Specifying a type will restrict the results to this type. If multiple types are specified, the API will return all addresses that match any of the types. Note: this parameter is available only for requests that include an API key or a client ID. If both result_type and location_type restrictions are present then the API will return only those results that matches both the result_type and the location_type restrictions.

latlng(data)

The latitude and longitude values specifying the location for which you wish to obtain the closest, human-readable address.

@tags [reverse]

location(data)

Returns a standartized location as a struct %{Geocoder.Location}.

place_id(data)

The place ID of the place for which you wish to obtain the human-readable address.

@tags [reverse]

query(data, type)

Builds a query parameters for the implementation.

result_type(data)

One or more address types.

@tags [reverse]